Tweak packaging script to autogenerate launchers
# Conflicts:
#	package.json


Former-commit-id: 160b81d4fc6cb55a7df1dd12cf44c68cb74be3e3 [formerly 1267737f0010b8973d43192e4cdd4e251ded1588] [formerly 6651871a1cf41d249b64db631c9eefccc73a461d [formerly d898a71cc8a9b3e788656e64e93716a598184663]]
Former-commit-id: 724a01a946e31b376e0d4cbd54100dd0868f9fe1 [formerly 0def88cf38bf3ef2133e8b6f920e0a544cb481f5]
Former-commit-id: 692ee25cb46bdba39ce829bdddf996ad7891177b

+++ scripts/Build_NWJS.ps1
@@ -6,20 +6,39 @@ $ZIMbase = "wikipedia_en_100_maxi"
 cd $PSScriptRoot
 cd ..
 foreach ($build in $builds) {
-    $target = "bld\nwjs\$build-$version\kiwix_js_windows-$appBuild"
-    $fullTarget = "$target-$build"
+    "Building $build $version..."
+    $OBuild = $build
     if ($build -eq "win-xp") {
         $build = "win-ia32"
         $version = $versionXP
     }
+    $target = "bld\nwjs\$build-$version\kiwix_js_windows-$appBuild"
+    $fullTarget = "$target-$build"
     $buildLocation = "node_modules\nwjs-builder-phoenix\caches\nwjs-v$version-$build.zip-extracted\nwjs-v$version-$build\"
     $archiveFolder = "$fullTarget\archives"
-    # Remove existing target
-    rm $fullTarget\* -Recurse
+    if (Test-Path $fullTarget -PathType container) {
+        "Removing directory $fullTarget..."
+        rm $fullTarget\* -Recurse
+        rm $fullTarget
+    }
     md $fullTarget
     # Copy latest binary x64
     cp $buildLocation\* $fullTarget -Recurse
     cp .\package.json, .\pwabuilder-sw.js, .\index.html, .\CHANGELOG.md, .\LICENSE, .\www $fullTarget -Recurse
+    "Copying archive..."
     md $archiveFolder
     cp "archives\$ZIMbase*.*", "archives\README.md" $archiveFolder
+    "Creating launchers..."
+    $launcherStub = ".\bld\nwjs\$build-$version\Start Kiwix JS Windows"
+    # Batch file
+    $batch = '@cd "' + "kiwix_js_windows-$appBuild-$build" + '"' + "`r`n" + '@start "Kiwix JS Windows" "nw.exe"' + "`r`n'"
+    $batch > "$launcherStub.bat"
+    # Shortcut
+    $WshShell = New-Object -ComObject WScript.Shell
+    $Shortcut = $WshShell.CreateShortcut("$launcherStub.lnk")
+    $Shortcut.TargetPath = '%windir%\explorer.exe'
+    $Shortcut.Arguments = "kiwix_js_windows-$appBuild-$build\nw.exe"
+    $Shortcut.IconLocation = '%windir%\explorer.exe,12'
+    $Shortcut.Save()
+    "Build $OBuild finished.`n"
 }
